ManualAnnotations, Saving
Manual Annotation settings are finished. In short, they now allow the user to toggle cells to the alternative category by double-clicking on them, or removing a label by control-double-clicking them (which only works in "only show labels" mode).

Saving now saves the current red index, the current manual annotations, and the feature cutoff values used to define the red index.

I fixed a bug that prevented all of the red feature histograms from updating across planes when one of the settings changed or when the data was loaded.