Handling session QC=False (++s2p helper)
Wrote function to check if session QC is correct (based on whether there is a scratch justification). It updates the database if requested.

Updated the "getTable" function to automatically ignore scratched sessions, unless explicitly told to include them.

Also, added the s2pTargets() function to needsS2P() so it's easy to start a new suite2p run on sessions that need it