added roiStackPosition, oneData handling
mpciROIs.stackPosition is now saved instead of planeIdx. It uses the weighted mean convention, which is now a top level method of vrExperiment. Some other code had to be changed to reflect this update.

vrRegistration can optionally clear previously saved oneData. This is mainly going to be used in pipelining, and not when adding or udpating specific oneData files.