fixed execute update statement +more
now instead of inserting a value into the execute statement, the default is to use the value as the second parameter of cursor.execute. This allows datetime and other datatypes to be inserted accurately.

Removed "time" as a second column because datetime objects can be read by python, even if they don't display correctly within access.

Added registrationDatestamp to the do registration method